Move each subcommand's parser into main_*
