+++ conda/cli/main_rename.py
@@ -7,11 +7,72 @@ Renames an existing environment by cloning it and then removing the original env
 from __future__ import annotations
 
 import os
-from argparse import ArgumentParser, Namespace
+from argparse import ArgumentParser, Namespace, _SubParsersAction
 from functools import partial
 from pathlib import Path
 
 from ..deprecations import deprecated
+
+
+def configure_parser(sub_parsers: _SubParsersAction, **kwargs) -> ArgumentParser:
+    from ..auxlib.ish import dals
+    from .helpers import add_parser_prefix
+
+    summary = "Rename an existing environment."
+    description = dals(
+        f"""
+        {help}
+
+        This command renames a conda environment via its name (-n/--name) or
+        its prefix (-p/--prefix).
+
+        The base environment and the currently-active environment cannot be renamed.
+        """
+    )
+    example = dals(
+        """
+        Examples::
+
+            conda rename -n test123 test321
+
+            conda rename --name test123 test321
+
+            conda rename -p path/to/test123 test321
+
+            conda rename --prefix path/to/test123 test321
+
+        """
+    )
+
+    p = sub_parsers.add_parser(
+        "rename",
+        help=summary,
+        description=description,
+        epilog=example,
+        **kwargs,
+    )
+    # Add name and prefix args
+    add_parser_prefix(p)
+
+    p.add_argument("destination", help="New name for the conda environment.")
+    # TODO: deprecate --force in favor of --yes
+    p.add_argument(
+        "--force",
+        help="Force rename of an environment.",
+        action="store_true",
+        default=False,
+    )
+    p.add_argument(
+        "-d",
+        "--dry-run",
+        help="Only display what would have been done by the current command, arguments, "
+        "and other flags.",
+        action="store_true",
+        default=False,
+    )
+    p.set_defaults(func="conda.cli.main_rename.execute")
+
+    return p
 
 
 @deprecated.argument("24.3", "24.9", "name")
@@ -52,7 +113,7 @@ def validate_destination(dest: str, force: bool = False) -> str:
     return dest
 
 
-def execute(args: Namespace, parser: ArgumentParser):
+def execute(args: Namespace, parser: ArgumentParser) -> int:
     """Executes the command for renaming an existing environment."""
     from ..base.constants import DRY_RUN_PREFIX
     from ..base.context import context
@@ -87,3 +148,4 @@ def execute(args: Namespace, parser: ArgumentParser):
             clone_and_remove()
     else:
         clone_and_remove()
+    return 0
