+++ conda/cli/main_run.py
@@ -6,8 +6,86 @@ Runs the provided command within the specified environment.
 """
 import os
 import sys
-from argparse import ArgumentParser, Namespace
+from argparse import REMAINDER, ArgumentParser, Namespace, _SubParsersAction
 from logging import getLogger
+
+
+def configure_parser(sub_parsers: _SubParsersAction, **kwargs) -> ArgumentParser:
+    from ..auxlib.ish import dals
+    from ..common.constants import NULL
+    from .actions import NullCountAction
+    from .helpers import add_parser_prefix, add_parser_verbose
+
+    summary = "Run an executable in a conda environment."
+    description = dals(
+        f"""
+        {summary}
+        """
+    )
+    example = dals(
+        """
+        Example::
+
+        $ conda create -y -n my-python-env python=3
+        $ conda run -n my-python-env python --version
+        """
+    )
+
+    p = sub_parsers.add_parser(
+        "run",
+        help=summary,
+        description=description,
+        epilog=example,
+        **kwargs,
+    )
+
+    add_parser_prefix(p)
+    add_parser_verbose(p)
+
+    p.add_argument(
+        "--dev",
+        action=NullCountAction,
+        help="Sets `CONDA_EXE` to `python -m conda`, assuming the current "
+        "working directory contains the root of conda development sources. "
+        "This is mainly for use during tests where we test new conda sources "
+        "against old Python versions.",
+        dest="dev",
+        default=NULL,
+    )
+
+    p.add_argument(
+        "--debug-wrapper-scripts",
+        action=NullCountAction,
+        help="When this is set, where implemented, the shell wrapper scripts"
+        "will use the echo command to print debugging information to "
+        "stderr (standard error).",
+        dest="debug_wrapper_scripts",
+        default=NULL,
+    )
+    p.add_argument(
+        "--cwd",
+        help="Current working directory for command to run in. Defaults to "
+        "the user's current working directory if no directory is specified.",
+        default=os.getcwd(),
+    )
+    p.add_argument(
+        "--no-capture-output",
+        "--live-stream",
+        action="store_true",
+        help="Don't capture stdout/stderr (standard out/standard error).",
+        default=False,
+    )
+
+    p.add_argument(
+        "executable_call",
+        nargs=REMAINDER,
+        help="Executable name, with additional arguments to be passed to the executable "
+        "on invocation.",
+    )
+
+    p.set_defaults(func="conda.cli.main_run.execute")
+
+    return p
 
 
 def execute(args: Namespace, parser: ArgumentParser) -> int:
